WIP: build_context: improve reproducibility by keeping soname ordered
While trakcing a reproducibility issue from maturin's output, we found
that the .so file in the output .whl were not ordered the same
everytime.

Order in the .whl comes down to the order soname_map is iterated.
HashMap does not provide a stable order. So switch to std::BTreeMap
which keep keys sorted and is iterable in natural order.

Signed-off-by: Yoann Congal <yoann.congal@smile.fr>